Move records out of views.py into their own file
