+++ core/views/__init__.py
@@ -3,3 +3,4 @@ from .validation_scenario import *
 from .organization import *
 from .record_group import *
 from .job import *
+from .record import *
